Data.Array.Accelerate.Math.DFT.Roots

Description

Documentation

rootsOfUnity :: (Shape sh, Slice sh, Floating e, FromIntegral Int e) => Exp (sh :. Int) -> Acc (Array (sh :. Int) (Complex e))

Calculate the roots of unity for the forward transform

inverseRootsOfUnity :: (Shape sh, Slice sh, Floating e, FromIntegral Int e) => Exp (sh :. Int) -> Acc (Array (sh :. Int) (Complex e))

Calculate the roots of unity for an inverse transform