Data.Array.Accelerate.Math.FFT

Description

For performance, compile against the foreign library bindings (using any number of '-fllvm-ptx', and '-fllvm-cpu' for the accelerate-llvm-ptx, and accelerate-llvm-native backends, respectively).

Documentation

data Mode

Constructors

Forward
Forward DFT

Reverse
Inverse DFT, un-normalised

Inverse
Inverse DFT, normalised

type FFTElt e = (Num e, RealFloat e, FromIntegral Int e, IsFloating e)

The type of supported FFT elements; namely Float and Double.

fft :: forall sh e. (Shape sh, Slice sh, FFTElt e) => Mode -> Acc (Array (sh :. Int) (Complex e)) -> Acc (Array (sh :. Int) (Complex e))

Discrete Fourier Transform along the innermost dimension of an array.

Notes for FFI implementations:

fftw supports arrays of dimension 1-5

cuFFT supports arrays of dimension 1-3

The pure implementation will be used otherwise.

fft1D :: forall e. FFTElt e => Mode -> Acc (Array DIM1 (Complex e)) -> Acc (Array DIM1 (Complex e))

Discrete Fourier Transform of a vector.

fft2D :: forall e. FFTElt e => Mode -> Acc (Array DIM2 (Complex e)) -> Acc (Array DIM2 (Complex e))

Discrete Fourier Transform of a matrix.

fft3D :: forall e. FFTElt e => Mode -> Acc (Array DIM3 (Complex e)) -> Acc (Array DIM3 (Complex e))

Discrete Fourier Transform of a 3D array.